Implement support popen for charliecloud launcher via bee_cluster class
