+++ bee-launcher/bee_cluster.py
@@ -1,6 +1,9 @@
 # system
-
+import os
+from subprocess import Popen, PIPE, \
+    STDOUT, CalledProcessError
 from threading import Thread, Event
+from termcolor import cprint
 
 
 class BeeTask(Thread):
@@ -63,3 +66,49 @@ class BeeTask(Thread):
         :param new_event:
         """
         self.__event_list.append(new_event)
+
+    # Task management support functions (private)
+    def run_popen_safe(self, command, nodes=None, shell=False):
+        """
+        Run defined command via Popen, try/except statements
+        built in and message output when appropriate
+        :param command: Command to be run
+        :param nodes: Defaults to os.environ['SLURM_NODELIST']
+                        Use to specify range of nodes message
+                        applies too (pass string!)
+        :param shell: Shell flag (boolean), default false
+        :return:
+        """
+        try:
+            p = Popen(command, shell, stdout=PIPE, stderr=STDOUT)
+            out, err = p.communicate()
+            if out:
+                print()
+                self.__handle_message(msg=out, nodes=nodes)
+            if err:
+                self.__handle_message(msg=err, nodes=nodes,
+                                      color=self.error_color)
+        except CalledProcessError as e:
+            self.__handle_message(msg="Error during - " + str(command) + "\n" +
+                                  str(e), nodes=nodes, color=self.error_color)
+        except OSError as e:
+            self.__handle_message(msg="Error during - " + str(command) + "\n" +
+                                  str(e), nodes=nodes,  color=self.error_color)
+
+    @staticmethod
+    def __handle_message(msg, nodes=None, color=None):
+        """
+        :param msg: To be printed to console
+        :param nodes: Defaults to os.environ['SLURM_NODELIST']
+                        Use to specify range of nodes message
+                        applies too (pass string!)
+        :param color: If message is be colored via termcolor
+                        Default = none (normal print)
+        """
+        list_nodes = nodes
+        if nodes is None:
+            list_nodes = os.environ['SLURM_NODELIST']
+        if color is None:
+            print(list_nodes + " " + msg)
+        else:
+            cprint(list_nodes + " " + msg, color)
